Se verifica que el total sea un entero

+++ dimabe_rrhh/models/wizard.py
@@ -89,7 +89,7 @@ class WizardHrPaySlip(models.TransientModel):
         payslips = self.env['hr.payslip'].sudo().search(
             [('indicadores_id', '=', indicadores.id), ('state', 'in', ['done', 'draft']),('employee_id.address_id.id','=',self.company_id.id)])
         totals = self.env['hr.payslip.line'].sudo().search([('slip_id', 'in', payslips.mapped('id'))]).filtered(
-            lambda a: a.total > 0 and type(a.total) == int)
+            lambda a: a.total > 0 and type(a.total) == float)
 
         totals_result = []
         for pay in payslips:
